requirements.txt
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6
celery==5.3.0
redis==4.5.5
python-telegram-bot==20.3
gunicorn==20.1.0
django-cors-headers==4.2.0